+++ include/afs/helper.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <cstdint>
+
 class Helper
 {
 public:
     static bool isFileExist(const char* filePath);
     static int openExistingFile(const char* filePath);
-    static size_t getCorrectSize(size_t inputSize); 
+    static uint32_t getCorrectSize(uint32_t inputSize);
 };
